requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
Deprecated==1.2.18
fastapi==0.115.8
filelock==3.17.0
flatbuffers==20181003210633
fsspec==2025.2.0
h11==0.14.0
humanfriendly==10.0
idna==3.10
Jinja2==3.1.5
loguru==0.7.3
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==2.8.8
numpy==2.0.2
onnxruntime==1.18.0
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==5.29.3
psutil==6.1.1
pydantic==2.10.6
pydantic_core==2.27.2
pyserial==3.5
python-dateutil==2.9.0.post0
python-multipart==0.0.20
python_tsp==0.5.0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
sympy==1.13.1
tabulate==0.8.10
torch==2.6.0
tsplib95==0.7.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
utils==1.0.2
uvicorn==0.34.0
wrapt==1.17.2
